Small Bug Fixes + Like/Comment Button
Like button is almost functional. The only problem left to figure out is determining how to mkae it so that the heart button is filled when a user has liked a post, and the heart button is unfilled when the user hasn't.
CURRENT IDEA:
Button image can easily be changed, but the button may need to be part of the actual PostCell, only way the current user ID can be checked for each post.